LayoutTests/fast/dom/DOMException/stack-trace-expected.txt

PASS "stack" in e is true
PASS typeof e.stack is "string"
PASS e.stack.contains("innerFunction") is true
PASS e.stack.contains("outerFunction") is true
PASS Object.prototype.toString.call(e) is "[object DOMException]"
PASS e.name is "HIERARCHY_REQUEST_ERR"
PASS e.message is "HIERARCHY_REQUEST_ERR: DOM Exception 3"
PASS e.code is 3
FAIL e.stack should be 42 (of type number). Was appendChild@[native code]
innerFunction@file:///src/WebKit/LayoutTests/fast/dom/DOMException/stack-trace.html:17
outerFunction@file:///src/WebKit/LayoutTests/fast/dom/DOMException/stack-trace.html:21
global code@file:///src/WebKit/LayoutTests/fast/dom/DOMException/stack-trace.html:27 (of type string).
PASS successfullyParsed is true

TEST COMPLETE
